+++ bugsnag-android-performance-api/src/main/kotlin/com/bugsnag/android/performance/EnabledMetrics.kt
@@ -1,7 +1,7 @@
 package com.bugsnag.android.performance
 
 /**
- * Sets which metrics (if any) are gathered and reported
+ * Sets which metrics (if any) are gathered and reported.
  */
 public class EnabledMetrics(
     /**