+++ bugsnag-android-performance/src/test/java/com/bugsnag/android/performance/test/NoopSpanProcessor.kt
@@ -1,0 +1,7 @@
+@file:Suppress("unused", "MatchingDeclarationName")
+
+package com.bugsnag.android.performance.test
+
+internal object NoopSpanProcessorPlaceholder
+
+internal const val NOOP_SPAN_PROCESSOR_MARKER: Int = 0