+++ bugsnag-android-performance/src/test/java/com/bugsnag/android/performance/test/TestTimeoutExecutor.kt
@@ -1,0 +1,7 @@
+@file:Suppress("unused", "MatchingDeclarationName")
+
+package com.bugsnag.android.performance.test
+
+internal object TestTimeoutExecutorPlaceholder
+
+internal const val TEST_TIMEOUT_EXECUTOR_MARKER: Int = 0