+++ features/fixtures/mazeracer/app/src/main/java/com/bugsnag/mazeracer/PerformanceTestUtils.java
@@ -4,6 +4,7 @@ import android.annotation.SuppressLint;
 
 import com.bugsnag.android.performance.internal.BugsnagPerformanceImpl;
 import com.bugsnag.android.performance.internal.InstrumentedAppState;
+import com.bugsnag.android.performance.internal.Persistence;
 import com.bugsnag.android.performance.internal.SpanProcessor;
 import com.bugsnag.android.performance.internal.processing.Tracer;
 
@@ -30,5 +31,14 @@ public class PerformanceTestUtils {
             Tracer tracer = (Tracer) processor;
             tracer.forceCurrentBatch();
         }
+    }
+
+    /**
+     * @noinspection KotlinInternalInJava
+     */
+    @SuppressLint("RestrictedApi")
+    public static void clearBugsnagPerformanceState() {
+        LogKt.log("PerformanceTestUtils.clearBugsnagPerformanceState()");
+        new Persistence(MazeRacerApplication.Companion.applicationContext()).clear();
     }
 }